working boost converter

+++ src/ckt/pwm.h
@@ -22,7 +22,7 @@ struct pwm {
 
 };
 
-void pwm_init(struct pwm *pwm, double dt, double D, int T, bool *out);
+void pwm_init(struct pwm *pwm, double dt, double D, double T, bool *out);
 void pwm_set_d(struct pwm *pwm, double D);
 void pwm_set_t(struct pwm *pwm, double T);
 void pwm_tick(struct pwm *pwm);
